+++ tests/bootstrap.php
@@ -31,14 +31,14 @@ if (!function_exists('createSomeFiles')) {
     function createSomeFiles(array $files = [])
     {
         $files = $files ?: [
+            TMP . 'exampleDir' . DS . '.hiddenDir' . DS . 'file7',
+            TMP . 'exampleDir' . DS . '.hiddenFile',
             TMP . 'exampleDir' . DS . 'file1',
             TMP . 'exampleDir' . DS . 'subDir1' . DS . 'file2',
             TMP . 'exampleDir' . DS . 'subDir1' . DS . 'file3',
             TMP . 'exampleDir' . DS . 'subDir2' . DS . 'file4',
             TMP . 'exampleDir' . DS . 'subDir2' . DS . 'file5',
             TMP . 'exampleDir' . DS . 'subDir2' . DS . 'subDir3' . DS . 'file6',
-            TMP . 'exampleDir' . DS . '.hiddenFile',
-            TMP . 'exampleDir' . DS . '.hiddenDir' . DS . 'file7',
         ];
 
         //Creates directories and files