+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 * updated for `php` 7.1 and `phpunit` 8.
 
 ## 1.2 branch
+### 1.2.8
+* `can_be_string()` renamed as `is_stringable()`. `can_be_string()` is
+    deprecated and it will be removed in a later release;
+* added `is_absolute()` global function;
+* uses `symfony/dom-crawler` for `BodyParser` class;
+* uses `symfony/filesystem` and `symfony/finder` for some global functions.
+
 ### 1.2.7
 * `getProperty()`, `getProperties()` and `invokeMethod()` methods from the
     `ReflectionTrait` can now take a class name as string, rather than just an